Introduce new field `verified_at`.
The new field will be added next to the current field `verified`, which stores the information if a user has
been verified its email address. Internally the new field is used for verification by check the field against `null`.

The new field makes it possible to not only store _if_ a user has been verified but also _when_.

Closes #177