Implement expiration of tokens
The current version does not support tokens who are expire. This could be done with adding a new field `expire_date` to
the user table right next to the `confirmation_token` field. But this pollutes the `users` table with an extra one-time-use field as the
`confirmation_token` already is.

Therefore lets create a new table which handles the confirmation tokens and the expiration date.