Fixing the case when a user is already verified
If the user is already verified, Laravel tries to bind the `token` parameter to the route. This will fail since the requested token was deleted while the user
was verified. To avoid this, the type hint has to be removed.

Furthermore we have to request the user first and check if he is already activated. If not, we request the user with the token and proceed with the token checks